moved all static methods to helpers script
